Refactoring the menu.
